+++ src/components/views/images/breed-modal/has-breed-content/index.ts
@@ -1,0 +1,3 @@
+import HasBreedContent from "./has-breed-content";
+
+export default HasBreedContent;
